Move filesystem.hpp to plugins
